+++ tests/cli/test_init_pandas.py
@@ -4,7 +4,6 @@ import shutil
 
 from click.testing import CliRunner
 
-from great_expectations import DataContext
 from great_expectations.cli import cli
 from great_expectations.data_context.util import file_relative_path
 from great_expectations.util import gen_directory_tree_str
@@ -165,99 +164,3 @@ great_expectations/
         raise e
     finally:
         os.chdir(curdir)
-
-
-def test_cli_init_connection_string_non_working_mssql_connection_instructs_user_and_leaves_entries_in_config_files_for_debugging(
-    tmp_path_factory,
-):
-    basedir = tmp_path_factory.mktemp("mssql_test")
-    basedir = str(basedir)
-    os.chdir(basedir)
-
-    runner = CliRunner()
-    result = runner.invoke(
-        cli,
-        ["init", "--no-view"],
-        input="Y\n2\n5\nmy_db\nmssql+pymssql://scott:tiger@not_a_real_host:1234/dbname\n",
-    )
-    stdout = result.output
-
-    assert "Always know what to expect from your data" in stdout
-    assert "What data would you like Great Expectations to connect to" in stdout
-    assert "Which database backend are you using" in stdout
-    assert "What is the url/connection string for the sqlalchemy connection" in stdout
-    assert "Give your new data source a short name" in stdout
-    assert "Attempting to connect to your database. This may take a moment" in stdout
-    assert "Cannot connect to the database" in stdout
-    assert "Database Error: No module named 'pymssql'" in stdout
-
-    assert "Profiling" not in stdout
-    assert "Building" not in stdout
-    assert "Data Docs" not in stdout
-    assert "Great Expectations is now set up" not in stdout
-
-    assert result.exit_code == 1
-
-    ge_dir = os.path.join(basedir, DataContext.GE_DIR)
-    assert os.path.isdir(ge_dir)
-    config_path = os.path.join(ge_dir, DataContext.GE_YML)
-    assert os.path.isfile(config_path)
-
-    # TODO this entry might not be totally right, but one needs to be here.
-    config = yaml.load(open(config_path, "r"))
-    assert config["datasources"] == {
-        "my_db": {
-            "data_asset_type": {
-                "module_name": None,
-                "class_name": "SqlAlchemyDataset",
-            },
-            "credentials": "${my_db}",
-            "class_name": "SqlAlchemyDatasource",
-            "generators": {
-                "default": {"class_name": "TableGenerator"},
-                "passthrough": {"class_name": "PassthroughGenerator"},
-            },
-        }
-    }
-
-    # TODO add entry in config_vars this entry might not be totally right, but one needs to be here.
-    config_path = os.path.join(
-        ge_dir, DataContext.GE_UNCOMMITTED_DIR, "config_variables.yml"
-    )
-    config = yaml.load(open(config_path, "r"))
-    assert config["my_db"] == {
-        "connection_string": "mssql+pymssql://scott:tiger@not_a_real_host:1234/dbname"
-    }
-
-    obs_tree = gen_directory_tree_str(os.path.join(basedir, "great_expectations"))
-    assert (
-        obs_tree
-        == """\
-great_expectations/
-    .gitignore
-    great_expectations.yml
-    datasources/
-    expectations/
-    notebooks/
-        pandas/
-            create_expectations.ipynb
-            validation_playground.ipynb
-        spark/
-            create_expectations.ipynb
-            validation_playground.ipynb
-        sql/
-            create_expectations.ipynb
-            validation_playground.ipynb
-    plugins/
-        custom_data_docs/
-            renderers/
-            styles/
-                data_docs_custom_styles.css
-            views/
-    uncommitted/
-        config_variables.yml
-        data_docs/
-        samples/
-        validations/
-"""
-    )
